# MRG, ENH: Speed up morph map creation