--- a/.gitignore
+++ b/.gitignore
@@ -44,6 +44,7 @@ build
 coverage
 .cache/
 .pytest_cache/
+__pycache__/
 prof/
 
 dist/