--- a/doc/whats_new.rst
+++ b/doc/whats_new.rst
@@ -34,7 +34,7 @@ Changelog
 
 - Accept filenames of raw .fif files that end in ``_meg.fif`` to enable complicance with the Brain Imaging Data Structure by `Stefan Appelhoff`_
 
-- Speed up morph map generation in :func:`mne.read_morph_maps` by ~5-10x by using :func:`numba.jit` by `Eric Larson`_
+- Speed up morph map generation in :func:`mne.read_morph_map` by ~5-10x by using :func:`numba.jit` by `Eric Larson`_
 
 - Add :class:`mne.digitization.Digitization` class to simplify montage by `Joan Massich`_