--- a/requirements.txt
+++ b/requirements.txt
@@ -7,6 +7,7 @@ pyqt5-sip
 sip
 scikit-learn
 nibabel
+numba
 h5py
 pandas
 numexpr