Speed up morph map generation ~~4x~~ ~10x using `numba` and changes to `_compute_nearest` calls. This code:
```
import time
import mne
subjects_dir = mne.datasets.sample.data_path() + '/subjects'
t0 = time.time()
mne.surface._make_morph_map('sample', 'fsaverage', xhemi=False,
                            subjects_dir=subjects_dir)
print(time.time() - t0)
```
- On `master` takes ~18 sec
- After refactoring to be a tiny bit faster takes 17 sec (what people will get without `numba` installed)
- On first run with numba takes ~~6~~ ~5 sec (includes compilation time)
- On subsequent runs takes ~~2.5~~ ~1.7 sec (no longer needs to compile)