@agramfort 2019-07-30
is the new code slower when numba is not available?