@larsoner 2019-07-31
On my macOS laptop, the same code snippet:

- `master`: 27.9 sec
- This PR without numba: 27.1 sec
- This PR with numba, first run: 11.2 sec
- This PR with numba, subsequent runs (because the compilation is cached): 5.7 sec