@larsoner 2019-07-31
... and 2.5 sec on my Linux machine now cut down to 1.76 sec by using cKDTree instead of BallTree in the nearest step.

I think I'll stop optimizing now :)