@agramfort 2019-08-03
ok I just tested. It works indeed much faster. 2 things:

- numba should be added to mne.sys_info() (note also that vtk + pyvista should be added too cc @GuillaumeFavelier )
- check that numba version is actually >= 0.4 on import?
- do we have 2 codes (with our without numba) tested in our CIs?

thx @larsoner !