@larsoner 2019-08-03
> do we have 2 codes (with our without numba) tested in our CIs?

Yes the `minimal` dependencies version should test it